+++ release.ps1
@@ -1,6 +1,6 @@
 param(
     [ValidatePattern('^\d+\.\d+\.\d+$')]
-    [string]$Version = "1.1.0",
+    [string]$Version = "1.2.0",
     [string]$DotNet = "dotnet"
 )